+++ app/index.js
@@ -51,7 +51,7 @@ const FriendCodeService = require('../db/friend_code_service.js');
 const MembersService = require('../db/members_service.js');
 const { variablesHandler } = require('./cmd/admin-cmd/environment_variables/variables_handler.js');
 const { createNewRecruitButton } = require('./buttons/recruit/components/create_recruit_buttons.js');
-const { handleCreateModal } = require('./modals/recruit/components/create_recruit_modal.js');
+const { handleCreateModal } = require('./modals/recruit/components/create_recruit_modals.js');
 const {
     modalRegularRecruit,
     modalAnarchyRecruit,
